Basic working dynamic site.
The meeting data files were moved as-is into the `data` directory.
Access to the meeting data is via a simple and completely unoptimised
`Data` class.

The Flask app now dynamically chooses the current month to display as
the main home content as well as listing future and past meetings.

Each meeting was now has a dedicated URL, allowing past meetings to be
viewed.